!14401 merge master into master

Created-by: yingguofeng
Commit-by: yingguofeng@huawei.com;openharmony_ci;zhanhao_;yangxiaoshuai2022;lukai25;milkpotatoes;p2003722;hecunmao;lvninglei;weihe99;guan-mingyue;。 。。。;w30079292;yaochaonan;starunvs;jiangmengyang;chentianyu;liJianfei;jinouwen;zhao1d;wuxiesaber;rentangyu;shao-yijiang;Geng Chen;sunzibo;Ilya Korobkov;caolili123;赵杜炜;zdd;xwcai98;chenlincl3;ZitongLi;hellowww321;gcw_rjxw9JkX;herongpeng;wangchen;zhangpeng611;panmingxuan99;chenjingxiang;gomader;dengwenjun;luobinghao;yejianghao;zhang_bo_xiao;wanghaha1204;tuxiaohang;杨云飞;chenwenrui;lijiangfeng;karenoni;zoumujia;liuhcunlin;tangjiao;shilei91;junzitang;liujia178;yangliping01;Like;dingding;like;jizaiyi;weixin_38992276;yanzhiqi1;dingwen;xing-yunhao-huawei;liuqian;tension16;zsc;yycc;zhangsaiyang1;zhengweiwei;xuziming;liu-zelin;cheryl2;yuxiaofei9;xiongluo;duyuhan;zh;Onlynagesha;l00809476;lijianfei;wwzm-jjw;shead-master;lijiamin2025;huangzhenghua;ningqicheng;Sunflower613;ohci1;hewei;ywx;xiaomujiang668;Naveen Sharma;huangtianzhi
Merged-by: ohci1
Description: ### 关联的Issue

### 修改原因（目的、解决的问题等，例如：修复xx场景崩溃问题）

### 修改描述（做了什么，变更了什么，例如：xx函数入口增加判空）

### 自测试项（测试结果截图，直接贴到每一个测试项底下）
#### jit-workload 本地执行（compiler/jit/isa/stub/ic/对象布局等修改必须执行）
- [ ] 已通过
- [ ] 不涉及，无需验证
#### interpreter-js_perf 性能测试（性能优化/性能敏感场景：IR化/ic/hclass/对象布局/isa/builtins等修改必须执行）
- [ ] 已验证不劣化
- [ ] 不涉及，无需验证
#### aot-js_perf 性能测试（性能优化/性能敏感场景：IR化/ic/hclass/对象布局/isa/builtins等修改必须执行）
- [ ] 已验证不劣化
- [ ] 不涉及，无需验证
#### interpreter-workload性能测试（性能优化/性能敏感场景：IR化/ic/hclass/对象布局/isa/builtins等修改必须执行）
- [ ] 已验证不劣化
- [ ] 不涉及，无需验证
#### aot-workload性能测试（性能优化/性能敏感场景：IR化/compiler/aot/isa/stub/ic/对象布局等修改必须执行）
- [ ] 已验证不劣化
- [ ] 不涉及，无需验证
#### 非StableArray不能采用ElementKind读取元素（修改解释器相关IR代码需要自检）
- [ ] 已自检
- [ ] 不涉及，无需验证
#### 在IR时添加fastpath时，需要检查是否有分支会改变非字典序模式或者修改数组长度，此时应该走slowpath（修改解释器相关IR代码需要自检）
- [ ] 已自检
- [ ] 不涉及，无需验证
#### IR化后的逻辑需要与c++实现一致，不一致的地方需要有分支用例看护（修改解释器相关IR代码需要自检）
- [ ] 已自检
- [ ] 不涉及，无需验证
#### c++代码中检测或抛出异常的地方，IR需要同样实现（修改解释器相关IR代码需要自检）
- [ ] 已自检
- [ ] 不涉及，无需验证
#### IR中使用默认入参是HeapObject对象的接口时，需要在上下文中保证了参数的正确性（修改解释器相关IR代码需要自检）
- [ ] 已自检
- [ ] 不涉及，无需验证
#### 修改c解释器，需要同时检查汇编解释器与AOT是否需要修改
- [ ] 已自检
- [ ] 不涉及，无需验证
### 是否已执行L0用例
- [ ] 已验证
- [ ] 不涉及。如不涉及，请写明理由
#### 修改或新增IR代码时，必须在代码注释中标注其对应的C++实现函数名（修改解释器相关IR代码需要自检）
‷
示例格式：
//@ref namespace::class::method()
‷
- [ ] 已自检
- [ ] 不涉及，无需验证
#### 单次代码提交的变更行数（新增+删除）超过200行时，必须在合并前组织线上Review
- [ ] 已自检
- [ ] 不涉及，无需验证

See merge request: openharmony/arkcompiler_ets_runtime!14412

+++ ecmascript/mem/c_string.h
@@ -128,7 +128,7 @@ inline std::string_view IntToStringView(int n, char* buffer, int bufferSize)
         n = -n;
         isNeg = false;
     }
-    size_t position = bufferSize;
+    size_t position = static_cast<size_t>(bufferSize);
     do {
         buffer[--position] = '0' - (n % DEC_BASE);
         n /= DEC_BASE;
@@ -136,7 +136,7 @@ inline std::string_view IntToStringView(int n, char* buffer, int bufferSize)
     if (isNeg) {
         buffer[--position] = '-';
     }
-    return {buffer + position, bufferSize - position};
+    return {buffer + position, static_cast<size_t>(bufferSize) - position};
 }
 
 template<class T>
